If version < 1 this is definitely not a iojs

+++ index.js
@@ -12,19 +12,19 @@ module.exports.safe = function() {
     }
     var isIo = _isIo;
 
-    if (!!~['v1.0.0','v1.0.1'].indexOf(version)) {
+    if (isIo && !~['v1.0.0','v1.0.1'].indexOf(version)) {
         var fs    = require('fs');
         var path  = require('path');
         var exec  = require('child_process').exec;
 
-        var temp  = path.join(process.cwd(), '.is-iojs-' + Math.random());
+        var temp  = path.join(process.cwd(), '.is-iojs-ey4Be6vDPrGfXGey4Be6vDPrGfXG');
         var epath = process.execPath;
 
         exec(epath + ' -h && echo "done" > ' + temp, function(err, help) {
             if (!err) isIo = /iojs/m.test(help);
         });
 
-        while (!(fs.existsSync(temp))) {}
+        while (!fs.existsSync(temp)) {}
         fs.unlinkSync(temp);
     }
